detect-tls-sni: move unittests to tests/
